KOrganizer's Birthdays: Fix leap year glitch when calculating bdays / anniversaries
svn path=/branches/KDE/4.4/kdepim/; revision=1128985

+++ korganizer/views/agendaview/koagendaitem.cpp
@@ -76,7 +76,7 @@ KOAgendaItem::KOAgendaItem( Calendar *calendar, Incidence *incidence,
 
   if ( mIncidence->customProperty( "KABC", "BIRTHDAY" ) == "YES" ||
        mIncidence->customProperty( "KABC", "ANNIVERSARY" ) == "YES" ) {
-    qint64 years = KOHelper::yearDiff( mIncidence->dtStart().date(), qd );
+    int years = KOHelper::yearDiff( mIncidence->dtStart().date(), qd );
     if ( years > 0 ) {
       mIncidence = incidence->clone();
       mIncidence->setReadOnly( false );
